finished server.py

+++ server.py
@@ -20,8 +20,8 @@ parser.add_argument("--command_file", type=str, default=None, help="filename to 
 parser.add_argument("--run_dir", type=str, default=None, help="location to run commands")
 parser.add_argument("--experiment_dir", type=str, default=None)
 
-parser.add_argument("--mem_gpu", type=int, default=1000, help="gpu memory needed for each job (in MB)")
-parser.add_argument("--mem_cpu", type=int, default=1000, help="cpu memory needed for each job (in MB)")
+parser.add_argument("--job_gpu_mem", type=int, default=0, help="gpu memory needed for each job (in MB)")
+parser.add_argument("--job_cpu_mem", type=int, default=0, help="cpu memory needed for each job (in MB)")
 parser.add_argument("--max_jobs_gpu", type=int, default=10, help="max number of jobs per gpu")
 parser.add_argument("--max_jobs_cpu", type=int, default=2, help="max number of jobs per cpu")
 
@@ -103,40 +103,49 @@ def create_experiment(args):
 def get_status_txt_new(metadata):
     status_strs = np.array([job["status_str"] for job in metadata["jobs"]])
     # status2str = {1000: "pending", 1001: "reserved", 1002: "running", 0: "finished", 1: "crashed", -2: "siginted", -9: "sigkilled", -15: "sigtermed"}
-    status_strs_unique = sorted(list(set(status_strs).union(set(status2str.values()))))
+    status_strs_unique = sorted(list(set(status_strs).union(set(status2str.values())).union({"unknown"})))
     nodes = np.array(["n/a" if job["node"] is None else job["node"] for job in metadata["jobs"]])
     nodes_unique = sorted(list(set(nodes).union({"n/a"})))
 
     d = np.zeros((len(nodes_unique), len(status_strs_unique)), dtype=object)
     for i_node, node in enumerate(nodes_unique):
         for i_status, status_str in enumerate(status_strs_unique):
-            d[i_node, i_status] = np.sum((nodes == node) & (status_strs == status_str))
-            if d[i_node, i_status] == 0:
-                d[i_node, i_status] = ""
+            a = np.sum((nodes == node) & (status_strs == status_str))
+            a = " " * 5 if a == 0 else f"{a: 5d}"
+            d[i_node, i_status] = a
     df = pd.DataFrame(d, index=nodes_unique, columns=status_strs_unique)
-    return tabulate(df, headers="keys", tablefmt="psql")
+    return tabulate(df, headers="keys", tablefmt="simple_grid")
 
 
 class Server:
     def __init__(self, args):
         self.args = args
-        self.get_server_resources()
+        self.metadata = None
+        self.this_node = os.uname().nodename
 
+        self.my_resources = self.get_server_resources()
         self.i_iter = 0
         self.popen2i_job = {}
 
     def get_server_resources(self):
-        args = self.args
         n_cpus = psutil.cpu_count()
         n_gpus = torch.cuda.device_count()
         mem_cpu = psutil.virtual_memory().available // 1000000
         mem_gpus = [torch.cuda.get_device_properties(i).total_memory // 1000000 for i in range(n_gpus)]
-        mem_gpus = [3000, 6000]
+
+        assert n_gpus > 0 or (n_gpus == 0 and self.args.job_gpu_mem == 0), "No GPUs found, and jobs require GPU memory."
+        if n_gpus == 0:
+            n_gpus = 1
+            mem_gpus = [1000]
+            self.args.max_jobs_gpu = 10000
+
+        self.args.mem_cpu = mem_cpu
+        self.args.mem_gpus = mem_gpus
 
         # gpu mem/time bound
-        n_procs_avail_per_gpu = [min(mem_gpu_i // args.mem_gpu, args.max_jobs_gpu) for mem_gpu_i in mem_gpus]
+        n_procs_avail_per_gpu = [min(mem_gpu_i // self.args.job_gpu_mem if self.args.job_gpu_mem > 0 else np.inf, self.args.max_jobs_gpu) for mem_gpu_i in mem_gpus]
         # cpu mem/time bound
-        n_procs_avail_cpu = min(mem_cpu // args.mem_cpu, args.max_jobs_cpu * n_cpus)
+        n_procs_avail_cpu = min(mem_cpu // self.args.job_cpu_mem if self.args.job_cpu_mem > 0 else np.inf, self.args.max_jobs_cpu * n_cpus)
         # cpu/gpu bound
         n_procs = min(n_procs_avail_cpu, sum(n_procs_avail_per_gpu))
         n_procs_per_gpu = [0 for _ in n_procs_avail_per_gpu]
@@ -144,63 +153,78 @@ class Server:
             for i_gpu in range(len(n_procs_avail_per_gpu)):
                 if n_procs_per_gpu[i_gpu] < n_procs_avail_per_gpu[i_gpu]:
                     n_procs_per_gpu[i_gpu] += 1
-        self.n_procs_per_gpu = n_procs_per_gpu
+        self.args.n_procs_per_gpu = n_procs_per_gpu
+        return dict(n_procs_per_gpu=n_procs_per_gpu)
+
+    def reserve_one_job(self, i_job, i_gpu):
+        with open(f"{self.args.experiment_dir}/{i_job:010d}/run.sh", "w") as f:
+            f.write("# !/bin/zsh\n")
+            f.write(f"echo $HOME\n")
+            f.write(f"source ~/.zshrc\n")
+            # f.write(f"source ~/activate_conda.sh\n")
+            f.write(f"conda activate {self.args.conda_env}\n")
+            f.write(f"cd {self.args.run_dir}\n")
+            f.write(f"export CUDA_VISIBLE_DEVICES={self.metadata['jobs'][i_job]['gpu']}\n")
+            f.write(f"{self.metadata['jobs'][i_job]['command']}\n")
+        self.metadata["jobs"][i_job]["status"] = 1001
+        self.metadata["jobs"][i_job]["status_str"] = status2str[1001]
+        self.metadata["jobs"][i_job]["node"] = self.this_node
+        self.metadata["jobs"][i_job]["gpu"] = i_gpu
 
     def reserve_jobs(self):
-        idxs_jobs = []
-        idxs_jobs = [i_job for i_job, job in enumerate(metadata["jobs"]) if job["status"] == 1000]
-        idxs_jobs = idxs_jobs[:3]
+        # pending jobs
+        idxs_jobs_pending = [i_job for i_job, job in enumerate(self.metadata["jobs"]) if job["status"] == 1000]
 
+        n_procs_running = [0 for _ in self.my_resources["n_procs_per_gpu"]]
+        running_popens = [popen for popen in self.popen2i_job.keys() if popen.poll() is None]
+        for popen in running_popens:
+            i_job = self.popen2i_job[popen]
+            n_procs_running[self.metadata["jobs"][i_job]["gpu"]] += 1
+        n_procs_avail = [a - b for a, b in zip(self.my_resources["n_procs_per_gpu"], n_procs_running)]
+        n_procs_launched = [0 for _ in n_procs_avail]
+
+        idxs_jobs = idxs_jobs_pending[: sum(n_procs_avail)]
+
+        i_gpu_abs = 0
         for i_job in idxs_jobs:
-            with open(f"{args.experiment_dir}/{i_job:010d}/run.sh", "w") as f:
-                f.write("# !/bin/zsh\n")
-                f.write(f"echo $HOME\n")
-                f.write(f"source ~/.zshrc\n")
-                # f.write(f"source ~/activate_conda.sh\n")
-                f.write(f"conda activate {args.conda_env}\n")
-                f.write(f"cd {args.run_dir}\n")
-                f.write(f"export CUDA_VISIBLE_DEVICES={metadata['jobs'][i_job]['gpu']}\n")
-                f.write(f"{metadata['jobs'][i_job]['command']}\n")
-            metadata["jobs"][i_job]["status"] = 1001
-            metadata["jobs"][i_job]["node"] = os.uname().nodename
+            for i_gpu_abs in range(i_gpu_abs, i_gpu_abs + len(n_procs_avail)):
+                i_gpu = i_gpu_abs % len(n_procs_avail)
+                if n_procs_launched[i_gpu] < n_procs_avail[i_gpu]:
+                    self.reserve_one_job(i_job, i_gpu)
+                    n_procs_launched[i_gpu] += 1
+                    break
         return idxs_jobs
-
-    def write_jobs(args):
-        pass
 
     def launch_jobs(self, idxs_jobs):
         for i_job in idxs_jobs:
-            job = metadata["jobs"][i_job]
+            job = self.metadata["jobs"][i_job]
             with open(f"{job['job_dir']}/out.txt", "w") as out_file, open(f"{job['job_dir']}/err.txt", "w") as err_file:
                 # start_new_session=True is needed so SIGINT doesn't propagate to child processes
                 popen = subprocess.Popen(f"zsh {job['job_dir']}/run.sh".split(" "), shell=False, start_new_session=True, stdout=out_file, stderr=err_file)
-            metadata["jobs"][i_job]["pid"] = popen.pid
-            popen2i_job[popen] = i_job
+            self.metadata["jobs"][i_job]["pid"] = popen.pid
+            self.popen2i_job[popen] = i_job
 
-    def run_server(args):
-        this_node = os.uname().nodename
-        i_iter = 0
+    def run_server(self):
+        quitted = False
         while True:
-            i_iter += 1
+            self.i_iter += 1
             try:
-                time.sleep(1)
-                with filelock.FileLock(f"{args.experiment_dir}/metadata.json.lock"):
-                    with open(f"{args.experiment_dir}/metadata.json", "r") as f:
-                        metadata = json.load(f)
+                time.sleep(2)
+                print("\x1b[2J\x1b[H", end="")
+                with filelock.FileLock(f"{self.args.experiment_dir}/metadata.json.lock"):
+                    with open(f"{self.args.experiment_dir}/metadata.json", "r") as f:
+                        self.metadata = json.load(f)
 
-                    # print("\r" * len(status_str.split("\n")), end="")
+                    idxs_jobs = self.reserve_jobs() if not quitted else []
+                    self.launch_jobs(idxs_jobs)
+                    # if len(idxs_jobs) > 0:
+                    # print(f"Launching {len(idxs_jobs)} jobs.")
 
-                    popen2poll = {popen: popen.poll() for popen in popen2i_job}
+                    popen2poll = {popen: popen.poll() for popen in self.popen2i_job}
                     popen2poll = {popen: (poll if poll is not None else 1002) for popen, poll in popen2poll.items()}
-
-                    idxs_jobs = reserve_jobs(args, metadata)
-                    launch_jobs(args, metadata, idxs_jobs, popen2i_job)
-                    if len(idxs_jobs) > 0:
-                        print(f"Launching {len(idxs_jobs)} jobs.")
-
                     for popen, poll in popen2poll.items():
-                        i_job = popen2i_job[popen]
-                        job = metadata["jobs"][i_job]
+                        i_job = self.popen2i_job[popen]
+                        job = self.metadata["jobs"][i_job]
                         job["status"] = poll
                         job["status_str"] = status2str[poll]
                         try:
@@ -208,17 +232,16 @@ class Server:
                         except psutil.NoSuchProcess:
                             job["pid_status"] = "NoSuchProcess"
 
-                    with open(f"{args.experiment_dir}/metadata.json", "w") as f:
-                        json.dump(metadata, f, indent=4)
+                    with open(f"{self.args.experiment_dir}/metadata.json", "w") as f:
+                        json.dump(self.metadata, f, indent=4)
 
-                    print("\x1b[2J\x1b[H", end="")
                     print(f"Running server!")
-                    print(f"Node: {this_node}")
+                    print(f"Node: {self.this_node}")
                     print(f"{datetime.now().strftime('%Y-%m-%d %H:%M:%S')}")
-                    print(f"Iteration {i_iter:010d}")
+                    print(f"Iteration {self.i_iter: 10d}")
                     print("Args:")
                     print("-" * 80)
-                    for k, v in vars(args).items():
+                    for k, v in vars(self.args).items():
                         print(f"{k:20s}: {v}")
                     print("-" * 80)
                     print()
@@ -231,26 +254,34 @@ class Server:
                     # idxs_jobs = [i_job for i_job, job in enumerate(metadata["jobs"]) if job["node"] == this_node]
                     # print(get_status_txt(metadata, idxs_jobs))
                     # print([metadata["jobs"][i_job]["pid_status"] for i_job in idxs_jobs])
-                    print(get_status_txt_new(metadata))
-
-                    no_jobs_pending = all([job["status"] != 1000 for job in metadata["jobs"]])
-                    my_jobs_done = all([poll != 1002 for poll in popen2poll.values()])
-                    if no_jobs_pending and my_jobs_done:
+                    print(get_status_txt_new(self.metadata))
+                    # no_jobs_pending = all([job["status"] != 1000 for job in self.metadata["jobs"]])
+                    # my_jobs_done = all([poll != 1002 for poll in popen2poll.values()])
+                    # print(no_jobs_pending, my_jobs_done, [popen.poll() for popen in popen2poll])
+                    # if no_jobs_pending and my_jobs_done:
+                    #     break
+                    should_wait = lambda status: status == 1002 or status == 1001 or (status == 1000 and not quitted)
+                    if not any([should_wait(job["status"]) for job in self.metadata["jobs"]]):
                         break
 
             except KeyboardInterrupt:
                 signal = input("Type signal to send to all jobs (ex. sigint=2, sigkill=9, sigterm=15): ")
-                signal = int(signal)
-                print("Sending signal to all jobs: ", signal)
-                for popen in popen2i_job:
-                    popen.send_signal(signal)
+                if signal.isdigit():
+                    signal = int(signal)
+                    print("Sending signal to all jobs: ", signal)
+                    for popen in self.popen2i_job:
+                        popen.send_signal(signal)
+                    quitted = True
+                else:
+                    print("Did not understand signal. Continuing...")
 
         print("Done! Exiting Server")
 
 
 def main(args):
     create_experiment(args)
-    run_server(args)
+    server = Server(args)
+    server.run_server()
 
 
 if __name__ == "__main__":
